fix error on serialization of metric_fu yaml report of .erb files
On Sonar Runner analysis of a Ruby project with .erb files, which
metric_fu puts in its report file the '!binary' type on the 'class_name'
and 'name' properties. The SnakeYAML library throws an error when
loading the yaml file string as a Map: see line 36 of
MetricfuComplexityYamlParserImpl class.

The fix is to replace, in the yaml file string, the string
'name: !binary' to the string 'name: tag:yaml.org,2002:binary'.